Abstract focal point handling to fn
